add wip for adding session participant id to the message model
removing the messenger id/type from message and using session participant id as foreign key to participant session. this is to allow a session with the same actor in it more than once so the logic can infer the current speaker using a unique participant id instead of the actor id.